chnge

+++ src/main/java/com/rob/devoloperops/backend/persistence/domain/backend/Plan.java
@@ -7,11 +7,11 @@ import javax.persistence.Id;
 @Entity
 public class Plan {
 
-    public int getId() {
+    public long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(long id) {
         this.id = id;
     }
 
@@ -27,7 +27,7 @@ public class Plan {
 
 
     @Id
-    private int id;
+    private long id;
 
     private String name;
